refactor transformer types
